Trait effectpedal::double_buf::DoubleBufPort

pub trait DoubleBufPort {
    type R;
    type W;

    fn rw<F, G>(&mut self, write: F, read: G)
    where
        F: Fn(&mut Self::W),
        G: FnMut(&Self::R);
}

One half of an asynchronous asymmetrical bidirectional double buffer. This can be either a Left or a Right.

Required Associated Types

type R

type W

Required Methods

fn rw<F, G>(&mut self, write: F, read: G) where
F: Fn(&mut Self::W),
G: FnMut(&Self::R),

Read and write to a DoubleBuf.

Arguments

‘write’ - Closure which writes data to the buffer. This will run every single time.

‘read’ - Closure which reads data from the buffer. This will not run if no fresh data is available!

When called on a Left, rw may briefly block waiting for the other side of the buffer to release. When called on a Right, rw should never block. No matter what, the write operation always happens first.